refactors for clearer scraping names.
