+++ lib/tiktok/getProfile.ts
@@ -2,7 +2,7 @@ import { Database } from "../../types/database.types";
 import getDataset from "../apify/getDataset";
 import getActorStatus from "../apify/getActorStatus";
 import getFormattedAccount from "./getFormattedAccount";
-import startProfileScraping from "./getProfileDatasetId";
+import startProfileScraping from "./startProfileScraping";
 
 type Social = Database["public"]["Tables"]["socials"]["Row"];
